Removed custom scrollbars from phones. Changed the way phone-specific markup works. Added locales and locales-related API. Refactored all the code to support locales. Moved SynthFlights-specific code from `js/layers` to `js/SynthFlightModules`.
