+++ js/LeafletAdvancedLayerSystem/Wizard.js
@@ -4,9 +4,9 @@
 L.ALS.Wizard = L.ALS.Widgetable.extend({
 
 	/**
-	 * Name of the layer type that will be displayed in the wizard window
+	 * Name of the layer type that will be displayed in the wizard window. You can use locale property to localize it.
 	 */
-	displayName: "Layer",
+	displayName: "layerWizardName",
 
 	initialize: function () {
 		L.ALS.Widgetable.prototype.initialize.call(this);
