+++ js/LeafletAdvancedLayerSystem/_service/WizardWindow.js
@@ -1,13 +1,13 @@
 L.ALS._service.WizardWindow = L.ALS._service.SidebarWindow.extend({
 	initialize: function (button) {
-		L.ALS._service.SidebarWindow.prototype.initialize.call(this, button, "Select layer type to add", "Set initial layer parameters");
+		L.ALS._service.SidebarWindow.prototype.initialize.call(this, button, "wizardSelectTitle", "wizardContentTitle");
 		this.container.id = "wizard-content";
 		this.select.id = "wizard-menu"
 
 		let addButton = document.createElement("div");
 		addButton.className = "button-base";
 		addButton.id = "wizard-add-button";
-		addButton.textContent = "Add";
+		L.ALS.Locales.localizeElement(addButton, "wizardAddButton");
 		this.window.querySelector("div[data-id='buttons-wrapper']").appendChild(addButton);
 		addButton.addEventListener("click", () => {
 			this.windowContainer.setAttribute("data-hidden", "1");
