+++ js/LeafletAdvancedLayerSystem/locales/English.js
@@ -1,0 +1,59 @@
+L.ALS.Locales["English"] = {
+
+	// Menu
+
+	menuSaveButton: "Save",
+	menuLoadButton: "Load",
+	menuExportButton: "Export",
+	menuSettingsButton: "Settings",
+
+	// Sidebar window
+
+	sidebarWindowCancelButton: "Cancel",
+
+	// Wizard
+
+	wizardSelectTitle: "Select layer type to add",
+	wizardContentTitle: "Set initial layer parameters",
+	wizardAddButton: "Add",
+
+	// Settings
+
+	settingsSelectTitle: "Settings sections",
+	settingsContentTitle: "Settings for selected section",
+	settingsApplyButton: "Apply and Close",
+	settingsExportButton: "Export Settings",
+	settingsImportButton: "Import Settings",
+	settingsRevertButton: "Revert back to default value",
+	settingsLoadingNotSupported: "Sorry, your browser doesn't support file loading. Please, update it.",
+	settingsImportError: "File that you try to load is not SynthFlight settings file",
+	settingsSavingNotSupported: "Settings can't be saved upon page refresh because your browser doesn't support it. Please, install any modern browser, so it won't happen.",
+	settingsAboutItem: "About",
+
+	// General settings
+
+	generalSettingsLanguage: "Language",
+	generalSettingsTheme: "Theme",
+	generalSettingsLightTheme: "Light",
+	generalSettingsDarkTheme: "Dark",
+	generalSettingsSystemTheme: "System",
+	generalSettingsRangeTheme: "Time range",
+
+	// System
+
+	systemProjectAlreadyOpen: "You already have an opened project. If you'll open another project, your changes won't be saved. Are you sure you wan't to load another one?",
+	systemProjectLoadingNotSupported: "Sorry, your browser doesn't support project loading. However, you still can create a new project, save it and open it later in a newer browser.",
+	systemConfirmDeletion: "Are you sure you want to delete this layer?",
+	systemNotProject: "File that you try to load is not SynthFlight project.",
+
+	// Layer
+
+	layerWizardName: "Layer",
+	layerDefaultName: "Generic Layer",
+
+	// File widget
+
+	fileNoFilesSelected: "No files selected. Click here to select some files.",
+	fileSelectedFile: "Selected files:",
+
+};
