+++ js/LeafletAdvancedLayerSystem/locales/Russian.js
@@ -1,0 +1,59 @@
+L.ALS.Locales["Русский"] = {
+
+	// Menu
+
+	menuSaveButton: "Сохранить",
+	menuLoadButton: "Загрузить",
+	menuExportButton: "Экспортировать",
+	menuSettingsButton: "Настройки",
+
+	// Sidebar window
+
+	sidebarWindowCancelButton: "Отмена",
+
+	// Wizard
+
+	wizardSelectTitle: "Выберите тип слоя для добавления",
+	wizardContentTitle: "Установите изначальные параметры типа слоя",
+	wizardAddButton: "Добавить",
+
+	// Settings
+
+	settingsSelectTitle: "Разделы настроек",
+	settingsContentTitle: "Настройки для выбранного раздела",
+	settingsApplyButton: "Применить и Закрыть",
+	settingsExportButton: "Экспортировать Настройки",
+	settingsImportButton: "Импортировать Настройки",
+	settingsRevertButton: "Вернуть к изначальному значению",
+	settingsLoadingNotSupported: "Извините, ваш браузер не поддерживает загрузку файлов. Пожалуйста, установите современный браузер, чтобы этого не происходило.",
+	settingsImportError: "Файл, который вы пытаетесь загрузить, не является файлом настроек SynthFlight",
+	settingsSavingNotSupported: "Настройки не сохранятся при перезагрузке страницы, поскольку ваш браузер это не поддерживает. Пожалуйста, установите современный браузер, чтобы этого не происходило.",
+	settingsAboutItem: "О программе",
+
+	// General settings
+
+	generalSettingsLanguage: "Язык",
+	generalSettingsTheme: "Тема",
+	generalSettingsLightTheme: "Светлая",
+	generalSettingsDarkTheme: "Темная",
+	generalSettingsSystemTheme: "Системная",
+	generalSettingsRangeTheme: "По временному промежутку",
+
+
+	// System
+
+	systemProjectAlreadyOpen: "У вас уже открыт проект. Если вы откроете другой проект, изменения в текущем проекте не будут сохранены. Вы уверены, что хотите открыть другой проект?",
+	systemProjectLoadingNotSupported: "Извините, ваш браузер не поддерживает загрузку проектов. Однако, вы можете создать новый проект, сохранить его и позже открыть в более современном браузере.",
+	systemConfirmDeletion: "Вы уверены, что хотите удалить этот слой?",
+	systemNotProject: "Файл, который вы пытаетесь загрузить, не является проектом SynthFlight",
+
+	// Layer
+
+	layerWizardName: "Слой",
+	layerDefaultName: "Абстрактный Слой",
+
+	// File widget
+
+	fileNoFilesSelected: "Никакие файлы не выбраны. Нажмите сюда, чтобы выбрать файлы",
+	fileSelectedFile: "Выбранные файлы:",
+};
