+++ js/SynthFlightModules/SynthGridSettings.js
@@ -9,31 +9,31 @@ L.ALS.SynthGridSettings = L.ALS.Settings.extend({
 	initialize: function () {
 		L.ALS.Settings.prototype.initialize.call(this);
 
-		let gridBorderColorWidget = new L.ALS.Widgets.Color("gridBorderColor", "Default grid border color");
+		let gridBorderColorWidget = new L.ALS.Widgets.Color("gridBorderColor", "defaultGridBorderColor");
 		gridBorderColorWidget.setAttributes({
 			value: this.gridBorderColor,
 			defaultValue: this.gridBorderColor
 		});
 
-		let gridFillColorWidget = new L.ALS.Widgets.Color("gridFillColor", "Default grid fill color");
+		let gridFillColorWidget = new L.ALS.Widgets.Color("gridFillColor", "defaultGridFillColor");
 		gridFillColorWidget.setAttributes({
 			value: this.gridFillColor,
 			defaultValue: this.gridFillColor
 		});
 
-		let meridiansColorWidget = new L.ALS.Widgets.Color("meridiansColor", "Paths by meridians default color");
+		let meridiansColorWidget = new L.ALS.Widgets.Color("meridiansColor", "defaultMeridiansColor");
 		meridiansColorWidget.setAttributes({
 			value: this.meridiansColor,
 			defaultValue: this.meridiansColor
 		});
 
-		let parallelsColorWidget = new L.ALS.Widgets.Color("parallelsColor", "Paths by parallels default color");
+		let parallelsColorWidget = new L.ALS.Widgets.Color("parallelsColor", "defaultParallelsColor");
 		parallelsColorWidget.setAttributes({
 			value: this.parallelsColor,
 			defaultValue: this.parallelsColor
 		});
 
-		let lineThicknessWidget = new L.ALS.Widgets.Number("lineThickness", "Default line thickness");
+		let lineThicknessWidget = new L.ALS.Widgets.Number("lineThickness", "defaultLineThickness");
 		lineThicknessWidget.setAttributes({
 			min: 1,
 			max: 20,
@@ -41,7 +41,7 @@ L.ALS.SynthGridSettings = L.ALS.Settings.extend({
 			defaultValue: this.lineThickness
 		});
 
-		let optimizationWidget = new L.ALS.Widgets.Number("gridHidingFactor", "Increase responsiveness by hiding grid at higher zoom levels. The higher the following value, the more responsive program will be");
+		let optimizationWidget = new L.ALS.Widgets.Number("gridHidingFactor", "gridHidingFactor");
 		optimizationWidget.setAttributes({
 			min: 1,
 			max: 10,
