+++ js/SynthFlightModules/SynthShapefileSettings.js
@@ -6,13 +6,13 @@ L.ALS.SynthShapefileSettings = L.ALS.Settings.extend({
 	initialize: function () {
 		L.ALS.Settings.prototype.initialize.call(this);
 
-		let fillColor = new L.ALS.Widgets.Color("fillColor", "Default fill color");
+		let fillColor = new L.ALS.Widgets.Color("fillColor", "shapefileDefaultFillColor");
 		fillColor.setAttributes({
 			value: this.fillColor,
 			defaultValue: this.fillColor
 		});
 
-		let borderColor = new L.ALS.Widgets.Color("borderColor", "Default border color");
+		let borderColor = new L.ALS.Widgets.Color("borderColor", "shapefileDefaultBorderColor");
 		borderColor.setAttributes({
 			value: this.borderColor,
 			defaultValue: this.borderColor
