+++ js/SynthFlightModules/locales/English.js
@@ -1,0 +1,87 @@
+L.ALS.Locales.addLocaleProperties("English", {
+
+	// SynthGridWizard
+
+	gridWizardDisplayName: "Grid Layer",
+	gridWizardNotification: `If zoom is too low, generated grid will be hidden. Please, zoom in to see it.
+	
+	To select a polygon, either click right mouse button (or tap and hold) or double-click (or double-tap) on it.`,
+
+	gridStandardScales: "Grid scale",
+	gridLngDistance: "Distance between parallels",
+	gridLatDistance: "Distance between meridians",
+
+	// SynthGridLayer
+
+	gridLayerDefaultName: "Grid Layer",
+	hidePolygonWidgets: "Hide widgets on the map",
+	hideNumbers: "Hide points' numbers on the map",
+	hidePathsConnections: "Hide paths' connections",
+	hidePathsByMeridians: "Hide paths by meridians",
+	hidePathsByParallels: "Hide paths by parallels",
+	lineThickness: "Line thickness",
+	gridBorderColor: "Border color",
+	gridFillColor: "Fill color",
+	meridiansColor: "Paths by meridians color",
+	parallelsColor: "Paths by parallels color",
+	airportLat: "Airport latitude",
+	airportLng: "Airport longitude",
+	aircraftSpeed: "Aircraft speed (km/h)",
+	imageScale: "Image scale denominator",
+	cameraWidth: "Camera width (px)",
+	cameraHeight: "Camera height (px)",
+	pixelWidth: "Pixel size (μm)",
+	overlayBetweenPaths: "Overlay between images from adjacent paths (%)",
+	overlayBetweenImages: "Overlay between images from the same path (%)",
+	focalLength: "Focal length (mm)",
+	lngPathsCount: "Paths count in one cell by parallels",
+	latPathsCount: "Paths count in one cell by meridians",
+	lngPathsLength: "Length of paths by parallels",
+	latPathsLength: "Length of paths by meridians",
+	lngFlightTime: "Flight time by parallels",
+	latFlightTime: "Flight time by meridians",
+	lngCellSizeInMeters: "Mean cell width",
+	latCellSizeInMeters: "Mean cell height",
+	selectedArea: "Selected area",
+	flightHeight: "Flight height (m)",
+	lx: "Image height, lx",
+	Lx: "Image height on the ground, Lx",
+	Bx: "Distance between photographing positions, Bx",
+	ly: "Image width, ly",
+	Ly: "Image width on the ground, Ly",
+	By: "Distance between adjacent paths, By",
+	GSI: "GSI",
+	IFOV: "IFOV",
+	GIFOV: "GIFOV",
+	FOV: "FOV",
+	GFOV: "GFOV",
+
+	// SynthGridSettings
+
+	defaultGridBorderColor: "Default grid border color",
+	defaultGridFillColor: "Default grid fill color",
+	defaultMeridiansColor: "Paths by meridians default color",
+	defaultParallelsColor: "Paths by parallels default color",
+	defaultLineThickness: "Default line thickness",
+	gridHidingFactor: "Increase responsiveness by hiding grid at higher zoom levels. The higher the following value, the more responsive program will be",
+
+	// SynthShapefileWizard
+
+	shapefileDisplayName: "Zipped Shapefile Layer",
+	zippedShapefile: "Zipped shapefile",
+
+	// SynthShapefileLayer
+
+	shapefileDefaultName: "Zipped Shapefile",
+	shapefileNoFeatures: "This shapefile doesn't contain any features so it won't be added",
+	shapefileBroken: "Extent of this shapefile is broken, layer won't be fully displayed on the map. Please, open it in your favourite GIS and fix the extent.",
+	shapefileNotValid: "This file is not valid zipped shapefile",
+	shapefileBorderColor: "Border color",
+	shapefileFillColor: "Fill color",
+
+	// Shapefile settings
+
+	shapefileDefaultFillColor: "Default fill color",
+	shapefileDefaultBorderColor: "Default border color",
+
+});
